support for macro release flag
any macro flags will be read at the beginning of each macro table. the
case for "no loop" at the end of a macro is no longer needed as the loop
flag will just store the last valid index. once the end flag is reached,
the loop index is loaded into the offset. in order to stimulate a hold
at the release index, we check if the offset is already at the release
index. if it is, we leave the offset untouched.