pulse 2 (channel 1) and reworked delay system
pulse 2 is now implemented with no channel mixing yet, so only pulse 1
is output. delay system has been reworked to allow varying changes in
speed via Fxx. delays are now calculated on a per row basis, and the
amount of frames is dependant on the current speed at that row. Sxx and
Gxx calculations have also been simplified. they are also filtered as
they are read. if the Gxx or Sxx effect is 0 or greater than the song
speed, they are ignored.