triangle volume macro fix
the triangle channel would get re-enabled when it should be muted due to
the volume macro routine. a check for whether it has already been muted
has been placed. this isn't a complete permanent solution, but should
work for pretty much any song. if a song uses a volume macro that turns
the channel on and off multiple times, then this would need a more
permanent fix.