pitch macro and Pxx changes
pitch macro and Pxx offsets are now done similarly to the other effects
that change pitch. a final total offset that represents the total change for
the timers is now stored. before, an offset was stored, but it had not
yet been multiplied by 11.125, so more calculations still had to be done
to the offsets everytime they were needed. there was also the case of
overflowing for the pitch macro as 1 byte was not enough to cover the
entire 2 byte timer range.