fixed Gxx and Sxx post
Gxx and Sxx post calculations were done in the pre routine. This used
the speed at the time of the pre routine, so if the speed had changed by
the time the post routine was reached, the post calculation used was
incorrect. Gxx and Sxx post calculations are done in the post routine
now.