various fixes and reworks to delays/Gxx/Sxx
decrement for frames is done before testing for no delays in channel
routines. song speed is used in full when calculating for new frame
delays. Gxx and Sxx can now retrigger the channel routines when an
invalid Gxx and Sxx was entered before and after speed changes, allowing
a second read of the next frame.